ccportal/requirements.txt
django-simple-menu==1.0.9
django-apptemplates==0.0.1
django-sockjs-tornado>=0.0.1